+++ backend/services/conversation.py
@@ -19,11 +19,13 @@ def create_conversation(data: ConversationCreate) -> Conversation:
         raise
 
 
-def get_conversation_by_id(convo_id: str) -> Conversation:
+def get_conversation_by_id(convo_id: str, user_id: str) -> Conversation:
     try:
         convo = Conversation.objects(id=convo_id).first()  # type: ignore
         if not convo:
             raise HTTPException(status_code=404, detail="Conversation not found")
+        if all(str(member.id) != user_id for member in convo.members):
+            raise HTTPException(status_code=403, detail="You are not a member of this conversation")
         return convo
     except HTTPException:
         raise
@@ -32,11 +34,13 @@ def get_conversation_by_id(convo_id: str) -> Conversation:
         raise
 
 
-def delete_conversation(convo_id: str):
+def delete_conversation(convo_id: str, user_id: str):
     try:
         convo = Conversation.objects(id=convo_id).first()  # type: ignore
         if not convo:
             raise HTTPException(status_code=404, detail="Conversation not found")
+        if all(str(member.id) != user_id for member in convo.members):
+            raise HTTPException(status_code=403, detail="You do not have the rights to delete this conversation")
         convo.delete()
     except HTTPException:
         raise
